# inputs/openapi.yaml
openapi: 3.0.3
info:
  title: VMware vRA REST API Server
  version: 0.1.0
  description: |
    REST API server for VMware vRealize Automation (vRA) operations.
    
    This server provides programmatic access to VMware vRA functionality including:
    - Authentication and token management
    - Catalog item browsing and requesting
    - Deployment management and monitoring
    
    The API follows standard REST conventions and returns JSON responses.
  contact:
    name: VMware vRA CLI
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /catalog/items:
    get:
      tags:
        - catalog
      summary: List catalog items
      description: |
        Retrieve a list of available catalog items from vRA.
        Results can be filtered by project and paginated.
      parameters:
        - name: project_id
          in: query
          description: Filter by project ID
          required: false
          schema:
            type: string
        - name: page_size
          in: query
          description: Number of items per page (1-2000)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 100
        - name: first_page_only
          in: query
          description: Fetch only the first page
          required: false
          schema:
            type: boolean
            default: false
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of catalog items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemsResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /catalog/items/{item_id}:
    get:
      tags:
        - catalog
      summary: Get catalog item details
      description: Retrieve detailed information about a specific catalog item
      parameters:
        - name: item_id
          in: path
          description: Catalog item ID
          required: true
          schema:
            type: string
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Catalog item details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogItemResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /catalog/items/{item_id}/schema:
    get:
      tags:
        - catalog
      summary: Get catalog item schema
      description: Retrieve the request schema for a specific catalog item
      parameters:
        - name: item_id
          in: path
          description: Catalog item ID
          required: true
          schema:
            type: string
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Catalog item schema
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogSchemaResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /catalog/items/{item_id}/request:
    post:
      tags:
        - catalog
      summary: Request catalog item
      description: |
        Submit a request for a catalog item deployment.
        This creates a new deployment from the specified catalog item.
      parameters:
        - name: item_id
          in: path
          description: Catalog item ID
          required: true
          schema:
            type: string
        - name: verbose
          in: query
          description: Enable verbose HTTP logging
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CatalogRequestRequest'
      responses:
        '200':
          description: Catalog item request submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogRequestResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Catalog item not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    BaseResponse:
      type: object
      properties:
        success:
          type: boolean
          default: true
          description: Indicates if the operation was successful
        message:
          type: string
          nullable: true
          description: Human-readable message about the operation
        timestamp:
          type: string
          format: date-time
          description: ISO timestamp of the response
      required:
        - success
        - timestamp
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            success:
              type: boolean
              default: false
            error_code:
              type: string
              nullable: true
              description: Machine-readable error code
            error_details:
              type: object
              nullable: true
              description: Additional error details
          required:
            - success
...
    CatalogItemsResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            items:
              type: array
              items:
                type: object
                description: Catalog item data
              description: List of catalog items
            total_count:
              type: integer
              description: Total number of items returned
            page_info:
              type: object
              nullable: true
              properties:
                page_size:
                  type: integer
                first_page_only:
                  type: boolean
                project_filter:
                  type: string
                  nullable: true
              description: Pagination and filtering information
          required:
            - items
            - total_count
    CatalogItemResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            item:
              type: object
              description: Detailed catalog item information
          required:
            - item
    CatalogSchemaResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            item_schema:
              type: object
              description: JSON schema defining the catalog item request structure
          required:
            - item_schema
    CatalogRequestRequest:
      type: object
      properties:
        project_id:
          type: string
          description: Target project ID for the deployment
        inputs:
          type: object
          nullable: true
          description: Input parameters for the catalog item
        reason:
          type: string
          nullable: true
          description: Business justification for the request
        name:
          type: string
          nullable: true
          description: Custom name for the deployment
      required:
        - project_id
      example:
        project_id: "12345678-1234-1234-1234-123456789012"
        inputs:
          vm_name: "web-server-01"
          cpu_count: 2
          memory_mb: 4096
        reason: "Development environment setup"
        name: "web-server-deployment"
    CatalogRequestResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            deployment_id:
              type: string
              nullable: true
              description: ID of the created deployment
            request_id:
              type: string
              nullable: true
              description: ID of the catalog request
          required:
            - deployment_id
            - request_id
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        VMware vRA access token obtained via the /auth/login endpoint.
        The token is automatically managed by the server after authentication.
security:
  - BearerAuth: []
...